pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "socialwarehouse"
version = "2.0.0-dev"
description = "Data warehouse and delta lake for geographic, demographic, and civic data — the DSTK replacement"
readme = "README.md"
license = {text = "Proprietary"}
requires-python = ">=3.11"
authors = [
    {name = "Dheeraj Chand", email = "dheeraj@siegeanalytics.com"},
]

dependencies = [
    "siege-utilities[geo]>=3.13.0",
    "Django>=5.2",
    "djangorestframework>=3.16",
    "celery[redis]>=5.6",
    "click>=8.1",
    "geopandas>=1.1",
    "pandas>=2.2",
    "pydantic>=2.10",
    "pydantic-settings>=2.0",
    "psycopg2-binary>=2.9",
    "sqlalchemy>=2.0",
    "geopy>=2.4",
]

[project.optional-dependencies]
spark = [
    "pyspark>=4.0",
    "delta-spark>=4.0",
    "apache-sedona>=1.7",
]
analysis = [
    "pysal>=24.1",
    "tobler>=0.12",
    "scipy>=1.14",
]
full = [
    "socialwarehouse[spark,analysis]",
    "polars>=1.0",
]
dev = [
    "ipython",
    "ipdb",
    "pytest>=8.0",
    "pytest-django>=4.9",
    "ruff>=0.9",
]

[project.scripts]
swh = "swh.cli:cli"

[tool.setuptools.packages.find]
include = ["swh*", "socialwarehouse*"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "socialwarehouse.settings.test"
pythonpath = ["."]
